first-class auth levels

+++ src/main.rs
@@ -26,7 +26,7 @@ async fn main() {
     let user = read_file("./experiment/username");
     let pass = read_file("./experiment/password");
     let auth_level = session.login(&user, &pass).await.unwrap();
-    println!("Auth level: {}", auth_level);
+    println!("Auth level: {}", auth_level as u8);
 
     /*
     let filedump = std::fs::read("experiment/test.torrent").unwrap();
